requirements.txt
pytest == 8.3.2
selene == 2.0.0rc9
allure-pytest == 2.13.5
requests == 2.32.3
python-dotenv == 1.0.1
Appium-Python-Client~=4.0.1